+++ example/src/main/java/ModelWithNoPackage.kt
@@ -1,0 +1,4 @@
+import io.github.newagewriter.mapper.Mapper
+
+@Mapper
+data class ModelWithNoPackage(val id: String, val name: String)